% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBAFphased-function.R
\name{getBAFphased}
\alias{getBAFphased}
\title{Function to use the solution of mixture modelling to phase the BAF}
\usage{
getBAFphased(x, seed = 1)
}
\arguments{
\item{x}{This is a list element created in the MiMMAl function.}

\item{seed}{This is the seed used when mirroring individual values. Default = 1.}
}
\description{
This function is primarily used within the MiMMAl function
}
\examples{
getBAFphased()
}
